Remove accidentally added old recipe implementation
This was an old method to remove custom serializers. A part of it got readded in 2caff96b86c52254912ab8e21b03b809f7ce47fa. Because only half was added it creates invalid packets when the code is hit. This would happen with mods like PolyPuck as this old method doesn't respect PolyMaps.